+++ src/shunting-yard.h
@@ -3,6 +3,8 @@
 // Use of this source code is governed by the BSD 2-Clause License that can be
 // found in the LICENSE file.
 
+#include <stdint.h>
+
 #ifndef SHUNTING_YARD_H
 #define SHUNTING_YARD_H
 
@@ -19,6 +21,6 @@ typedef enum {
 } Status;
 
 // Calculates the result of a mathematical expression.
-Status shunting_yard(const char *expression, double *result);
+Status shunting_yard(const char *expression, uint64_t *result);
 
 #endif  // SHUNTING_YARD_H
